# orchestrator/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Weave Orchestrator API
  version: 0.1.0
servers:
  - url: http://localhost:9999
    description: Local development
security:
  - ApiKeyAuth: []
...
paths:
...
  /tenants/{slug}/workflows:
    get:
      tags:
        - workflow
      summary: List workflows
      operationId: listWorkflows
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Workflows list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowResource"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - workflow
      summary: Save a workflow
      operationId: saveWorkflow
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowResource"
      responses:
        "200":
          description: Saved workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowResource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /tenants/{slug}/workflows/{workflow_id}:
    get:
      tags:
        - workflow
      summary: Get a workflow
      operationId: getWorkflow
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowResource"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - workflow
      summary: Delete a workflow
      operationId: deleteWorkflow
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
        - $ref: "#/components/parameters/TenantId"
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: >
        Tenant API secret. Send as Authorization Bearer token, or pass the same value in the X-API-Key header.
  parameters:
    TenantId:
      in: path
      name: slug
      required: true
      schema:
        type: string
...
    WorkflowId:
      in: path
      name: workflow_id
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: API key does not match tenant in path or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Plan quota exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [detail]
      properties:
        detail:
          type: string
      example:
        detail: resource not found
...
    WorkflowNodeResource:
      type: object
      additionalProperties: false
      required: [id, type, agent_id]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [agent]
          default: agent
        agent_id:
          type: string
        objective:
          type: string
          nullable: true
    WorkflowEdgeResource:
      type: object
      additionalProperties: false
      required: [from_node]
      properties:
        from_node:
          type: string
          description: Source node id or START
        to_nodes:
          type: array
          items:
            type: string
          default: []
        routes:
          type: object
          additionalProperties:
            type: string
          nullable: true
          description: Conditional route map (route value to target node id)
    WorkflowResource:
      type: object
      additionalProperties: false
      required: [id, name, description, nodes, edges]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowNodeResource"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowEdgeResource"
      example:
        id: ppl_log_analysis
        name: PPL Log RCA
        description: Generate PPL query, fetch logs, produce RCA report, and identify code issues
        nodes:
          - id: ppl_generation
            type: agent
            agent_id: ppl_generation
            objective: Generate PPL query
          - id: fetch_and_analyze
            type: agent
            agent_id: fetch_and_analyze
            objective: Fetch logs and produce RCA report
          - id: code_analysis
            type: agent
            agent_id: git_inference
            objective: Use log RCA findings to identify code issues in the repository
        edges:
          - from_node: START
            to_nodes: [ppl_generation, fetch_and_analyze, code_analysis]
...